Переименовывает функцию
